# Conflicts:
#	mzmine-community/src/main/java/io/github/mzmine/datamodel/features/types/annotations/CompoundNameType.java
#	mzmine-community/src/main/java/io/github/mzmine/datamodel/features/types/annotations/SpectralLibraryMatchesType.java
#	mzmine-community/src/main/java/io/github/mzmine/modules/dataprocessing/id_lipidid/common/identification/matched_levels/MatchedLipid.java

+++ mzmine-community/src/main/java/io/github/mzmine/datamodel/features/types/annotations/CompoundDatabaseMatchesType.java
@@ -183,7 +183,7 @@ public class CompoundDatabaseMatchesType extends ListWithSubsType<CompoundDBAnno
   }
 
   @Override
-  public int getPrefColumnWidth() {
+  public double getPrefColumnWidth() {
     return 150;
   }
 }
